Generate flow-typed stubs
Generated flow-typed stubs with `flow-typed install` to remove the flow
errors. Since those libraries miss some flow types, it was giving false
negatives and showing a ton of errors. It would be a good idea to add
libdefs for the libraries that we use frequently in the project in order
to have Flow's intellisense and type checking safety.