Add url-parse for the login
Since the default Node's url module is not yet supported (according to
the react-native red box message), I think url-parse is the best
alternative. It has almost the same functionnalities, but since the
module name is not the same, Flow doesn't mistakenly think it's the
original Node module. I also created flow libdefs for url-parse, and I
will remove them when my PR on the flow-typed repo has been merged.