Refactor and add 100% test coverage
I was behind on tests because I was busy making the prototype work. Now
that I knew how feasible the initial idea was, I caught up on tests to
make the code as solid as it can be. I also refactored the code to make
it easier to unit test and easier to separate into distinct modules.