Change the tests folder structure
It just makes more sense to have the unit tests right next to their
tested files, but I didn't know jest could do that in recent versions. I
simply removed the `__tests__` folders and brought the tests down a
level. It's now easier to find the tests for a file since it's right
next to it with the same name and the extension `.test`.