Remove the prepush hook
The prepush hook is useless when you have a precommit hook since you
can't push files that haven't been committed. Also, if you make changes
to a file that is in a non-pushed commit and the file doesn't validate,
the prepush hook will block you from pushing your commit, even though
the "bad" changes wouldn't get pushed. It really slows down the
iteration cycle as we add more tests and more files to the project.